pyproject.toml
[tool.poetry]
name = "kattis-cli"
version = "1.1.2"
authors = ["Ram Basnet <rbasnet@coloradomesa.edu>"]
description = "A command-line tool for Kattis"
readme = "README.md"
repository = "https://github.com/rambasnet/kattis-cli"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

# PEP 621 metadata table for tools that expect [project].
[project]
name = "kattis-cli"
version = "1.1.2"
description = "A command-line tool for Kattis"
readme = "README.md"
authors = [
    { name = "Ram Basnet", email = "rbasnet@coloradomesa.edu" }
]

[project.urls]
"Homepage" = "https://github.com/rambasnet/kattis-cli"
"Bug Tracker" = "https://github.com/rambasnet/kattis-cli/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.poetry.dependencies]
python = "^3.10.0"
requests = "^2.31.0"
click = "^8.1.7"
beautifulsoup4 = "^4.12.2"
pyyaml = "^6.0.1"
rich = "^13.6.0"
tomlkit = "^0.12.2"
lxml = "^6.0.1"
trogon = "^0.6.0"

[tool.poetry.scripts]
kattis = 'kattis_cli.main:main'

